Refactor `override_config_table/utilities.py`
