+++ tests/golden_config_infra/test_config_reload_with_rendered_golden_config.py
@@ -5,7 +5,7 @@ import os
 
 from tests.common.helpers.assertions import pytest_assert
 from tests.common.config_reload import config_reload, config_reload_minigraph_with_rendered_golden_config_override
-from tests.override_config_table.utilities import backup_config, restore_config, get_running_config, \
+from tests.common.utilities import backup_config, restore_config, get_running_config, \
     compare_dicts_ignore_list_order, NON_USER_CONFIG_TABLES
 from tests.common.utilities import update_pfcwd_default_state
