+++ tests/route/test_forced_mgmt_route.py
@@ -6,7 +6,7 @@ import pytest
 from tests.common.config_reload import config_reload
 from tests.common.helpers.assertions import pytest_assert
 from tests.common.utilities import wait_until
-from tests.override_config_table.utilities import backup_config, restore_config, \
+from tests.common.utilities import backup_config, restore_config, \
         reload_minigraph_with_golden_config
 from tests.syslog.syslog_utils import is_mgmt_vrf_enabled
